+++ tests/test_commandline.py
@@ -1,73 +1,136 @@
+import sys
 import os.path
 import shutil
+import logging
 import dicognito.__main__
 from .data_for_tests import load_dcm
 
 
-def test_overwrite_files(capsys):
-    orig_dataset = read_file(test_overwrite_files.__name__, "orig_p01_s01_s01_i01.dcm")
+def setup_module(module):
+    base_dir = os.path.dirname(__file__)
+    data_dir = os.path.join(base_dir, "data")
+    orig_dir = os.path.join(base_dir, "orig_data")
+    if os.path.isdir(data_dir):
+        shutil.rmtree(data_dir)
+    shutil.copytree(orig_dir, data_dir)
+
+
+def test_overwrite_files():
+    test_name = get_test_name()
+    orig_dataset = read_original_file(test_name, "p01_s01_s01_i01.dcm")
     assert "CompressedSamples^MR1" == orig_dataset.PatientName
 
-    copy_file(test_overwrite_files.__name__, "orig_p01_s01_s01_i01.dcm", "p01_s01_s01_i01.dcm")
-    run_dicognito(test_overwrite_files.__name__, capsys)
+    run_dicognito(path_to("p*"))
 
-    anon_dataset = read_file(test_overwrite_files.__name__, "p01_s01_s01_i01.dcm")
+    anon_dataset = read_file(test_name, "p01_s01_s01_i01.dcm")
     assert anon_dataset.PatientName != orig_dataset.PatientName
 
 
+def test_ignores_file_that_do_not_match_glob():
+    test_name = get_test_name()
+    orig_dataset = read_original_file(test_name, "np01_s01_s01_i01.dcm")
+    assert "CompressedSamples^MR1" == orig_dataset.PatientName
+
+    run_dicognito(path_to("p*"))
+
+    anon_dataset = read_file(test_name, "np01_s01_s01_i01.dcm")
+    assert anon_dataset.PatientName == orig_dataset.PatientName
+
+
 def test_summary_mixed_files_reports_on_each_study(capsys):
-    expected = """\
+    expected_output = """\
 Accession Number Patient ID       Patient Name
 ---------------- ----------       ------------
-821CFB2XY4A8     XRP26N7QWFM3     BYERS^YOLANDA^MYLES
-0ID4RQFGHICH     YKQBAJZS5UO3     MARSHALL^BEULAH^AURELIA
-NO4NP4PXEOWT     YKQBAJZS5UO3     MARSHALL^BEULAH^AURELIA
+DRVN05NEDUYD     2S183ZNON7HU     RICHMOND^MARCY^NITA
+8NZGNEJWE7QA     NPC1XHSJT51Z     MORROW^SUSANNA^LUCIEN
+SXJXM4HE90EO     NPC1XHSJT51Z     MORROW^SUSANNA^LUCIEN
 """
+    run_dicognito(path_to("p*"))
+    (actual_output, actual_error) = capsys.readouterr()
 
-    copy_file(
-        test_summary_mixed_files_reports_on_each_study.__name__, "orig_p01_s01_s01_i01.dcm", "p01_s01_s01_i01.dcm"
-    )
-    copy_file(
-        test_summary_mixed_files_reports_on_each_study.__name__, "orig_p01_s01_s01_i02.dcm", "p01_s01_s01_i02.dcm"
-    )
-    copy_file(
-        test_summary_mixed_files_reports_on_each_study.__name__, "orig_p02_s01_s01_i01.dcm", "p02_s01_s01_i01.dcm"
-    )
-    copy_file(
-        test_summary_mixed_files_reports_on_each_study.__name__, "orig_p02_s02_s01_i01.dcm", "p02_s02_s01_i01.dcm"
-    )
-
-    actual = run_dicognito(test_summary_mixed_files_reports_on_each_study.__name__, capsys)
-
-    assert expected == actual
+    assert expected_output == actual_output
 
 
 def test_summary_with_quiet_no_report(capsys):
-    copy_file(test_summary_with_quiet_no_report.__name__, "orig_p01_s01_s01_i01.dcm", "p01_s01_s01_i01.dcm")
-    copy_file(test_summary_with_quiet_no_report.__name__, "orig_p01_s01_s01_i02.dcm", "p01_s01_s01_i02.dcm")
-    copy_file(test_summary_with_quiet_no_report.__name__, "orig_p02_s01_s01_i01.dcm", "p02_s01_s01_i01.dcm")
-    copy_file(test_summary_with_quiet_no_report.__name__, "orig_p02_s01_s01_i01.dcm", "p02_s01_s01_i01.dcm")
+    expected_output = ""
 
-    actual = run_dicognito(test_summary_with_quiet_no_report.__name__, capsys, "--quiet")
-    expected = ""
-    assert expected == actual
+    run_dicognito(path_to("p*"), "--quiet")
+    (actual_output, actual_error) = capsys.readouterr()
+
+    assert expected_output == actual_output
 
 
-def run_dicognito(test_name, capsys, *extra_args):
+def test_directory_is_recursed():
+    test_name = get_test_name()
+    orig_dataset1 = read_original_file(test_name, "p01_s01_s01_i01.dcm")
+    orig_dataset2 = read_original_file(test_name, "a", "b", "p01_s02_s01_i01.dcm")
+    assert "CompressedSamples^MR1" == orig_dataset1.PatientName
+    assert "CompressedSamples^MR1" == orig_dataset2.PatientName
+
+    run_dicognito(path_to(""))
+
+    anon_dataset1 = read_file(test_name, "p01_s01_s01_i01.dcm")
+    anon_dataset2 = read_file(test_name, "a", "b", "p01_s02_s01_i01.dcm")
+    assert orig_dataset1.PatientName != anon_dataset1.PatientName
+    assert orig_dataset2.PatientName != anon_dataset2.PatientName
+
+
+def test_non_dicom_files_ignored(capsys):
+    expected_error = ""
+
+    test_name = get_test_name()
+    orig_dataset = read_original_file(test_name, "p01_s01_s01_i01.dcm")
+    assert "CompressedSamples^MR1" == orig_dataset.PatientName
+
+    run_dicognito(path_to(""))
+    (actual_output, actual_error) = capsys.readouterr()
+
+    anon_dataset = read_file(test_name, "p01_s01_s01_i01.dcm")
+    assert orig_dataset.PatientName != anon_dataset.PatientName
+    assert expected_error == actual_error
+
+
+def test_non_dicom_files_logged_at_info(caplog):
+    expected_error = "not_a_dicom_file.txt appears not to be DICOM. Skipping."
+
+    test_name = get_test_name()
+    orig_dataset = read_original_file(test_name, "p01_s01_s01_i01.dcm")
+    assert "CompressedSamples^MR1" == orig_dataset.PatientName
+
+    # py.test configures the logs itself, so setting the log level in the command
+    # doesn't work. Instead, use caplog to set the level.
+    caplog.set_level(logging.INFO)
+    run_dicognito(path_to(""))
+
+    anon_dataset = read_file(test_name, "p01_s01_s01_i01.dcm")
+    assert orig_dataset.PatientName != anon_dataset.PatientName
+
+    log_record = caplog.records[0]
+    assert log_record.levelname == "INFO"
+    assert log_record.getMessage().endswith(expected_error)
+
+
+def get_test_name():
+    depth = 1
+    while True:
+        frame = sys._getframe(depth)
+        if frame.f_code.co_name.startswith("test"):
+            return frame.f_code.co_name
+        depth += 1
+
+
+def path_to(end_of_path):
     base_dir = os.path.dirname(__file__)
-    glob = os.path.join(base_dir, "data", test_name, "p*.dcm")
-    dicognito.__main__.main(("--salt", test_name, glob) + extra_args)
-    (out, error) = capsys.readouterr()
-    return out
+    return os.path.join(base_dir, "data", get_test_name(), end_of_path)
 
 
-def read_file(directory, name):
-    return load_dcm(os.path.join("data", directory, name))
+def run_dicognito(*extra_args):
+    dicognito.__main__.main(("--salt", "salt for test") + extra_args)
 
 
-def copy_file(test_name, original_name, new_name):
-    base_dir = os.path.dirname(__file__)
-    original_path = os.path.join(base_dir, "data", test_name, original_name)
-    new_path = os.path.join(base_dir, "data", test_name, new_name)
+def read_file(*directory_parts):
+    return load_dcm(os.path.join("data", *directory_parts))
 
-    shutil.copyfile(original_path, new_path)
+
+def read_original_file(*directory_parts):
+    return load_dcm(os.path.join("orig_data", *directory_parts))